In wiktionary lookups, also try the cap'd word.
That is, try the capitalized word. For example, "paris" fails but
"Paris" succeeds. This is helpful since the unigram data set I have is
all-lowercase, and it's not always obvious how to capitalize the words.
For example, some unigrams like "md" are probably most used as "MD",
while others like "javascript" may be considered most correct when
written as "JavaScript." So this fix, while useful, does not fully
address the capitalization question.