Evals work even if an LLM match has a bad index.
Well, they work as well as they can in that situation. The thing is, if
a match index is incorrect, it's harder to trust that there is a correct
match. This will still count the match as correct, and will add an error
message in the browser as well as a more extensive error message in the
terminal running the server.