Minor browser-detection update.
I did this because I was trying to improve mouse wheel behavior on
Firefox. However, I found that if I dynamically turn on page snapping in
Firefox, it doesn't immediately snap to a page. I decided that the
default behavior in Firefox is good enough for now. It works a little
better than Chrome in that it doesn't get stuck (with the mouse wheel),
so I think it's ok after all.